Map AccountingDocument to Invoice

+++ legacy-cloud-acl/src/main/java/ro/linic/cloud/repository/AccountingDocumentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.query.Param;
 
 import ro.linic.cloud.entity.AccountingDocument;
 
-public interface AccountingDocumentRepository extends CrudRepository<AccountingDocument, Integer> {
+public interface AccountingDocumentRepository extends CrudRepository<AccountingDocument, Long> {
 	List<AccountingDocument> findByCompanyIdAndNrDoc(@Param("companyId") int companyId, @Param("nrDoc") String nrDoc);
 }
