+++ legacy-cloud-acl/src/main/java/ro/linic/cloud/repository/OperatiuneRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ro.linic.cloud.entity.Operatiune;
 
-public interface OperatiuneRepository extends CrudRepository<Operatiune, Integer> {
+public interface OperatiuneRepository extends CrudRepository<Operatiune, Long> {
 
 }
